stakes: polish collector_overrides
- remove ele_cnt (test only method)
- optimize get_root_idx
